requirements.txt
pytest~=8.3.3
selenium==4.25.0
allure-pytest~=2.13.5
sqlalchemy ~= 2.0.36
